Fixin.ai/requirements.txt
streamlit
Pillow
diffusers
transformers
google-generativeai
